+++ api/internal/order/model.go
@@ -10,7 +10,7 @@ type Order struct {
 	ID          uuid.UUID
 	UserID      uuid.UUID
 	Status      string
-	TotalAmount float64
+	TotalAmount int64
 	CreatedAt   time.Time
 }
 
@@ -18,7 +18,7 @@ type OrderItem struct {
 	ID        uuid.UUID
 	OrderID   uuid.UUID
 	ProductID uuid.UUID
-	Price     float64
+	Price     int64
 	Qty       int
 	CreatedAt time.Time
 }
